[Refactor] 파일 위치, 이름 수정
