+++ backend/timebank/src/main/kotlin/kookmin/software/capstone2023/timebank/presentation/api/v1/BankAccountController.kt
@@ -1,7 +1,7 @@
 package kookmin.software.capstone2023.timebank.presentation.controller
 
-import kookmin.software.capstone2023.timebank.application.service.bank.BankAccountCreateService
-import kookmin.software.capstone2023.timebank.application.service.bank.BankAccountReadService
+import kookmin.software.capstone2023.timebank.application.service.bank.account.BankAccountCreateService
+import kookmin.software.capstone2023.timebank.application.service.bank.account.BankAccountReadService
 import kookmin.software.capstone2023.timebank.presentation.api.RequestAttributes
 import kookmin.software.capstone2023.timebank.presentation.api.auth.model.UserAuthentication
 import kookmin.software.capstone2023.timebank.presentation.api.auth.model.UserContext
